Extract other classes out of Checklist
