+++ src/Main/Home.js
@@ -1,18 +1,16 @@
 import React, { Component } from 'react';
 
-import Wrapper from 'common/Wrapper';
 import ScrollFilledIcon from 'Icons/ScrollFilled';
 import MegaphoneIcon from 'Icons/Megaphone';
+import Wrapper from 'common/Wrapper';
+import DiscordButton from 'Main/DiscordButton';
 
 import ChangelogPanel from './ChangelogPanel';
-
 import MasteryRadiusImage from './Images/mastery-radius.png';
 import AntorusImage from './Images/antorus.jpg';
-
 import DiscordLogo from './Images/Discord-Logo+Wordmark-White.svg';
 import DiscordBotGif from './Images/discord-bot.gif';
 import SpecListing from './SpecListing';
-import DiscordButton from 'Main/DiscordButton';
 
 class Home extends Component {
   render() {
